Added ability to infer eta from data
If eta needs to be fixed then setting the standard deviation for the corresponding gaussian proposal distribution to 0 will produce the same results.

I've also added the ability to specify a prior on eta. Note priors on eta, beta and the partition are assumed to be independent at the moment.