Added information levels
RunPartitionsMCMC now takes an arguement which decides how much information is encoded in the intitial partition C0.

Low info: Only the total cases, contacts and number of households are known

Medium info: The number of cases, contacts and households disaggregated by household size is known.

High info: The full partition is known and C0 doesn't change in the fitting process.