Seperated out functions for different information levels and precalc logfactorials
The functions for selecting indices, moving contacts and calculating reverse are now separated out for low and medium information datasets. The functions SelectIndices and MoveContact are now obsolete but have been left in for now.

I have also changed the code so that the log factorials are all calculated at the start of the fit and are looked up as and when they are needed, speeding up the algorithm significantly.