Fixed final size solver for high transmission rates
When there are high transmission rates, the condition number of the matrix used to find final size distributions becomes very large making the inversion unstable. This previously gave an error message and returned inaccurate distributions. The condition number of the matrix is now checked before inversion and in the case where it is too large, we approximate the final size distribution as being all zeros except 1 at full population infection.